feat(vapix)!: Consolidate protocol helpers
Consolidating these in a common module makes it clear that they share a purpose that is distinct from other modules in the crate. One downstream benefit of this is that it becomes easier to see when their interfaces needlessly diverge.

Protocol helpers are in my current model of the crate one of several pillars:
- HTTP request and response representations - facilitate decoupling API binding implementations from the client and enable cassette testing.
- Client traits - facilitate decoupling API binding implementations from a specific client
- Client implementations - provide a battery included experience and deal with some annoying aspects such as binding host and credentials, and digest authentication
- Client independent API bindings (currently but not necessarily coupled with the trait) - document the existence and behavior of APIs in terms of Rust.
- Protocol helpers in three layers - facilitate implementation of API bindings inside and outside the crate.

This commit also:
- Improves the documentation in these modules to further clarify their purpose.
- Makes some items that should have been public public.

It stops short of moving the async functions because I'm not sure where to move them and I don't want code to churn needlessly.
