+++ crates/vapix/src/config/discover.rs
@@ -10,7 +10,10 @@ use semver::Version;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-use crate::http::{Error, HttpClient, Request};
+use crate::{
+    http::{HttpClient, Request},
+    protocol_helpers::http::Error,
+};
 
 const PATH: &str = "config/discover";
