+++ crates/vapix/src/config/recording_group_1.rs
@@ -6,8 +6,8 @@ use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 
 use crate::{
-    http::{Error, HttpClient, Request},
-    rest, rest_http,
+    http::{HttpClient, Request},
+    protocol_helpers::{http::Error, rest, rest_http},
 };
 
 #[derive(Debug, Deserialize, Serialize)]
