+++ crates/vapix/src/config/remote_object_storage_1_beta.rs
@@ -7,8 +7,8 @@ use serde_json::json;
 use url::Url;
 
 use crate::{
-    http::{Error, HttpClient, Request},
-    rest, rest_http,
+    http::{HttpClient, Request},
+    protocol_helpers::{http::Error, rest, rest_http},
 };
 
 const BASE_PATH: &str = "config/rest/remote-object-storage/v1beta/destinations";
