+++ crates/vapix/src/config/ssh_1.rs
@@ -9,8 +9,8 @@ use serde::{Deserialize, Serialize};
 use serde_json::json;
 
 use crate::{
-    http::{Error, HttpClient, Request},
-    rest, rest_http,
+    http::{HttpClient, Request},
+    protocol_helpers::{http::Error, rest, rest_http},
 };
 
 #[derive(Serialize)]
