+++ crates/vapix/src/protocol_helpers/soap.rs
